Fail tests if `@NNN` shows up in an error message
Summary: D81898793 fixes an issue in which variable ids were accidentally leaking into error messages, causing them to change between pyrefly runs. This isn't the first time I've made this mistake, so this diff adds a check to verify that we don't see anything that looks like a variable id in error mesages in tests.

Reviewed By: ndmitchell

Differential Revision: D81898791

fbshipit-source-id: b6f8bf3f77c4d57b79bdbca6debbe68cd7624aee