Add config so definition typeshed tests work without interpreter installed (#1057)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/1057

A bug report came in that Pyrefly doesn't build in `nix` when there's no Python interpreter available. Turns out we need it for some go-to-def functionality testing, but it isn't actually needed for things to work/Pyrefly to build.

[Original report](https://discord.com/channels/1319086885024567347/1319086885024567350/1414470301470953573)

This diff adds a config and moves some files around to get it working without an interpreter.

Reviewed By: grievejia

Differential Revision: D81943027

fbshipit-source-id: fbb63972b12f86aca256fdcc7e0f3156d922f263