Delete use-untyped-imports config option
Summary:
This diff deletes the `use-untyped-imports` flag, which shouldn't be used by anything anymore.

I've had an error being output for the past few weeks (months?) that should have warned people to migrate, so I think we might be good to go at this point. I also have been suspecting that it's a flag that wasn't ever used by anyone other than me, so I think we're fine to delete.

Reviewed By: grievejia

Differential Revision: D81944175

fbshipit-source-id: 5ebfcd53c23f6b60cf05ec2730f52f27016e2b12