basic support for TypeAliasType
Summary:
This is basically undocumented (only one off-hand mention in the typing spec with no examples).

It's what `type` statements generate at runtime, but it can also be used directly as a functional definition to produce an alias.

https://docs.python.org/3/library/typing.html#typing.TypeAliasType

This diff implements the basic support minus type params. We cannot reuse the scoped type alias binding, because type parameters are specified in a different way. The next diff implements type parameter support for TypeAliasType.

Reviewed By: ndmitchell

Differential Revision: D70990921

fbshipit-source-id: 5750623f5f37648c1bba06d73be8e3c4cd23876a