Add handler for the TSP "typeServer/getSupportedProtocol" request (#1029)
Summary:
This is the next stage in adding support for [Type Server Protocol](https://github.com/microsoft/pylance-release/discussions/7180) to Pyrefly

This is a very simple request that just returns the version number supported. It also includes the infrastructure for wiring up this request and testing it.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1029

Reviewed By: grievejia

Differential Revision: D81808996

Pulled By: kinto0

fbshipit-source-id: 242b4627f35752a08d9a6f55250892e58f9aab90