add buck target for tsp_types
Summary:
this was made by rich (not @ meta) and he did not use buck. buck is needed for the next diff where we put it in pyrefly

note: this does actually change `pyrefly/crates/tsp_types/Cargo.toml` quite a bit (but it shows up as generated and is hidden in review). if the tests pass, that means it's probably OK

Reviewed By: ndmitchell, migeed-z

Differential Revision: D81972926

fbshipit-source-id: 57bb58a8c213ee1ef3634603fb8f91b8b99ebfa7