Directly set the static info definition style
Summary:
I was skimming the scope code to understand the changes I'm going
to need to make for loop handling, and found what looked like a bug - we're
setting a completely incorrect definition style in the static info `add`
functions.

It turns out that it's not a bug, but our code is both wasetful of compute
and very confusing - we're manually creating a junk `style`, then returning
a mutable ref so that we can overwrite it immediately.

Let's clean this up - just pass the style in

Reviewed By: rchen152

Differential Revision: D82040321

fbshipit-source-id: 70be62bff7d3672b2334076b82033ec5a798ce36