Switch "Delete file" and "Add new file" buttons
Summary:
It feels more natural for the delete button to be immediately to the right of the filename and the add button to be further to the right. (That's how, e.g., tabs in Chrome are laid out.)

This also switches the order of the save and delete buttons when adding a new file, so I renamed the button to "Cancel" in this case so that it's obvious that it's a way to cancel the current file addition and not a lost delete button for a different file.

Reviewed By: stroxler

Differential Revision: D82084860

fbshipit-source-id: 287d62fe659b523f43da098908da471dfa4db709