handle unions in non-callable `__bool__` check
Summary:
When this was first implemented in D75118659 we skipped unions because as_call_target did not handle unions.

That was added a month or so later in D77437437, so i believe the special handling is no longer necessary.

The error message needs work, but we can handle that later. The problem is that we look up the attribute and get a single union type, so we lose track of which part of the base type's union maps to which part of the attribute type's union.

Reviewed By: ndmitchell

Differential Revision: D82112153

fbshipit-source-id: 712b93312814f363a96ca4f9f1622026a12365f2