Make is_subset_eq_var return a Result rather than a bool so we can include a failure reason
Summary:
I'm working on https://github.com/facebook/pyrefly/issues/817, which asks for a different error code for a bad override caused by mismatched function parameter names. I thought of two ways to do this:
1. Add a boolean parameter to is_subset_eq, called something like `check_param_names`, which controls whether we check that parameter names match. When a bad override is found, check whether it is still a bad override with check_param_names=false.
2. Change is_subset_eq to return a reason when two types are found to be incompatible. When a bad override is found, check whether the reason is mismatched parameter names.

(1) is easier to implement, but I'm going with (2) because it's something I've wanted to do for a while anyways, for better error messages.

This diff converts is_subset_eq_var to return a failure reason. Subsequent diffs will convert the remaining is_subset_eq helpers.

This conversion is a big, messy change, so I've introduced only one failure reason ("Other") and made liberal use of helpers to convert back and forth between bool and Result to keep the diffs as small as possible. Once the conversion is done, I'll clean up the code.

Reviewed By: ndmitchell

Differential Revision: D82169573

fbshipit-source-id: c4f57124ab9aee6e47532c9a2984a52415091413