Revert D81681451 (temporarily)
Summary:
The change in D81681451 is essentially correct, in Python most nested
scopes cannot see a containing class body.

But there are two bugs in our handling: comprehension iterators are being
evaluated in the wrong scope right now, and parameter scopes behave
differently (so parameter defaults are permitted to see the class body scope).

I suspect that the static-type-info carve out in this implementation is also
incorrect - if the variables from a class body are invisible, they aren't
usable as static types (trying to do so would give a name error and crash the
program, unless `__future__.annotations` is imported, and even if it is the
scoping rules shouldn't differ).

My current guess is that the static-type-info carve out was actually there
to approximate the parameter scope rule, but because defaults (which are not
static types) can *also* see the class body we want to implement it differently.

This diff backs out the change for now because with these issues present it
is doing more harm than good (we have zero bug reports about the original
behavior but 3 about the new behavior).

Reviewed By: ndmitchell

Differential Revision: D82227101

fbshipit-source-id: cd71e0bb5dd99e83542db81f6a867ed58a35748c