Add pydantic documentation
Summary:
Add basic documentations for some of the pydantic features we support.

NOTE:
A big todo is that we would like to add some  pydantic stubs into a sandbox and show some of these examples on it.

There are currently 5k lines of code as stubs. I will see if I can create some examples in sandbox so that the examples here can be viewed in sandbox instead.

Reviewed By: rchen152

Differential Revision: D82065121

fbshipit-source-id: fbd3a3155a3b177f7365b1ef99c278d4afba8fea