Stop converting back and forth between bool and Result
Summary:
In my previous diffs, I converted is_subset_* methods from returning bool to returning Result while changing as little code as possible, making liberal use of SubsetError::if_false and is_ok() as needed. This results in a lot of clunky code that throws away the SubsetError all the time by converting it back to a bool.

This diff adds three helpers - ok_or(), all(), and any() - and uses them to rewrite the code in a way that both is more readable and preserves the SubsetError.

Reviewed By: stroxler

Differential Revision: D82169580

fbshipit-source-id: fc62d3521a718808d22263b685fd87f6238a9803