Go back to path for module imports
Summary:
With the discussion on how to handle multiple target owners, we've decided to go with the single-target owner approach, meaning that for a given `ModulePath`, we will deterministically select one target to represent it.

Because of that, we actually don't need `Target` to exist in `Handle` anymore, or for handles to be the result of `ModuleName` lookups, so we can undo parts of that work.

`SourceDatabase`s still need to know about `Handle`s for metadata stuff, so we'll keep `Handle` in `pyrefly_build` for now, and maybe move it to `pyrefly_python` in the future.

Reviewed By: grievejia

Differential Revision: D82123716

fbshipit-source-id: ec5a047e40f844e8f04e300ad25840a09e4ea68b