Tweak the inherited efficiency hack a bit
Summary:
The class field struct ClassFieldInner is pretty big, and any new field in it
adds to the difficulty of understanding the actual semantics.

We added `name_might_exist_in_inherited` recently when breaking a cycle where
a class like
```
class A(list[A.B]):
    class B: pass
```
would have an MRO that depended on a class field, which caused problems when the
class field calculation tried to do an override consistency check. So we split
the override consistency check into a separate binding and passed a bit in order
to avoid unnecessarily scanning the MRO twice in certain cases where the class
field already has looked.

This is all fine, but we lacked a high-level comment (a comment was buried inside
of the monstrous `calculate_class_field`, but that wasn't in a place where you'd
see it if you were trying to understand what this part of the data structure is needed),
and also `name_might_exist_in_inherited` is pretty verbose, we can just use `is_inherited`.

This diff breaks that bit out of the `ClassFieldInner` data structure and describes
what's going on in a comment on the data types so that it's more discoverable.

Reviewed By: yangdanny97

Differential Revision: D82135323

fbshipit-source-id: 2a31b22ba34c7e11761e7be784dae0715f157c8f