Tweak the `style` update logic in `definitions.rs`
Summary:
When `style` stopped implementing `copy`, we introduced a `let style =`
binding and added a clone, but this isn't necessary, we can just clone.

This is important because the existing code is shadowing the function
argument `style` with the let binding, which means that on a skim it
looks like we're just overwriting the style, which would lead to
completely different semantics.

I got confused by this - when we skim code, we don't skim line-by-line
so the shadowing really is confusing. If we just directly clone, then
it's immediately obvious what the semantics actually are.

Reviewed By: rchen152

Differential Revision: D82325999

fbshipit-source-id: 2bdab6362fa2df8c9a44655384bd9a69b53eead1