Track some more edge cases
Summary:
After looking in depth at scope and name lookup logic, I realized that
while the current class body scope lookup rules look sensible there
are actually a few remaining issues.

The issues are edge cases and not all that important per se, but I have to
refactor this code in order to make changes I want for flow-merging to make
sense, and it's important to use the right primitives in the refactor.

Reviewed By: fangyi-zhou

Differential Revision: D82351923

fbshipit-source-id: 4530640673a9c24b0b8d0582bc48b00b7e23e3c2