The outermost comprehension iterator isn't in comprehension scope
Summary:
This diff closes a gap in the logic where class body scope elements
aren't visible in most nested scopes - we had to make an exception for
comprehensions in the initial version because too many use cases
relied on iterating over a class body variable, which works because
in the Python runtime the leftmost iterator is evaluated in the parent
scope.

Now, we delay creating the comprehension scope until after evaluating the
leftmost iterator, and as a result we can enforce the scope rule in
comprehensions (which fixes a bug and also simplfies our name lookup logic a
bit).

Reviewed By: fangyi-zhou

Differential Revision: D82349162

fbshipit-source-id: 5481dc25bb86801b03766489a1d13cf9819e6d02