Use depth rather than mutation for current scope
Summary:
In the previous diff, I caught some edge cases in our lookup logic; one of the
problems is that we need to special case not just the current scope but one scope
above the current scope to handle annotation scopes inside classes correctly.

To prep for that, I think it's nice to first express the current logic in terms
of scope depth rather than using a mutable `is_current_scope` bool.

Reviewed By: fangyi-zhou

Differential Revision: D82351922

fbshipit-source-id: 6c823e6a001f65227d5f52bf066238bf9ad6065c