Use the scope kind, not flow style, for class body logic
Summary:
A name defined in a class body cannot be seen except in specific circumstances.

The relevant question is whether it comes from class body scope, not whether
it has flow kind `ClassField` (which is a little misleading - that flow kind
is somewhat misnamed, it's a hack to get expressions into dataclass field
handling logic and only applies ot assignments, there are lots of class fields
that are defined in other ways).

This fixes the bug where a name defined in some way other than assignment (e.g.
a method definition) was bypassing the lookup rules.

Reviewed By: fangyi-zhou

Differential Revision: D82353435

fbshipit-source-id: c908ec75e8992753474ae03a917d09c16c62ea2d