Document a bug in name lookup logic
Summary:
Pyrefly is unable to detect an undefined local if the name exists
in parent scopes.

I found this bug because I'm looking at refactors of scope
to improve our flow-merging logic, and one of the options involves
unifying static and flow.

This would be benefical in ways that go beyond my immediate focus on flow
merging because the double-lookup leads to the actual case analysis not being
expressed clerly in rust types, which is part of the root cause of at least a
half dozen bugs (including this one) and also makes requires us using panics in
name lookups (because there are several places where we update flow, assume the
name *must* exist in static, and have to panic if it doesn't).

Reviewed By: ndmitchell

Differential Revision: D82378106

fbshipit-source-id: eb2e2e8e12cd78f90502aa1f5abdf22a340ca300