Document a mutable capture bug in class fields
Summary:
A mutable capture isn't actually in the local scope, and it won't
behave as though it were for the purposes of class field resolution.

We currently get this wrong, because `definitions` detects the write
to a global as a definition, and we don't model scope well enough to
later realize this was a mutable capture.

Reviewed By: ndmitchell

Differential Revision: D82378105

fbshipit-source-id: 617ca19dd3d690c4b79bf17b3a32754b914a8537