Distinguish between `Any` as a value and as a type
Summary:
Before this diff, we always converted the `Any` class def into `type[Any]`, but that makes it impossible to distinguish between `Any` as a value and `type[Any]`.

This diff removes the eager transformation, so `Any` as a value has the type `Type::ClassDef`. When we see this type in an annotation, we convert it to `Type::Any`.

One place we try to distinguish between `Any` as a value and `type[Any]` is `isinstance` checks, where `Any` as the 2nd argument is an error, but `type[Any]` is allowed, since it could have a materialization which is valid.

Another distinction is when `Any` (the value) is used as an attribute base or call target. Before this diff, we would allow any attribute on `Any`, but after this diff we error if you try to  access attributes that are not defined on `type`, matching other type checkers.

Type checkers also special case `Any` to forbid calling its constructor. This diff also implements this restriction.

Reviewed By: stroxler

Differential Revision: D82349368

fbshipit-source-id: dda8d1874214c52cc113d93c37be159fdca29d65