Track another mutable capture case that Python won't run
Summary:
We currently only approximate Python's rule that you can't use a
parent scope name that is a mutable capture prior to declaring
it incorrectly - we catch it if the function tries to *assign* the
capture prior to declaring it, but the actual restriction is that you
cannot use it at all, even for reads.

Tracking this because it will help inform refactors of scope that
I need for other reasons.

Reviewed By: ndmitchell

Differential Revision: D82378805

fbshipit-source-id: 78a1f0648eefdda543a867266241a89e9223ea04