Track a bug in ininitialized local handling
Summary:
While looking at refactors of scope that I want to improve narrowing
and flow merging (and loops), I realized that the current duplication
of data and reads is incorrect in uninitialized local detection - the
root cause is tied to the fact that storing static and flow info
separately makes it hard to always read them consistently.

Here, I track a bug where I realized the detection of initialization
is incorrectly using static lookups in a way where we might use the
flow to actually read, but the static info of an enclosing scope to
decide the read is initialized.

Reviewed By: ndmitchell

Differential Revision: D82384260

fbshipit-source-id: 89cc003541db5721779a597e8f50da56d75d22e5