Attach deprecated flags during auto complete (#1093)
Summary:
We know whether a method or a function is deprecated in Pyrefly, but the information is not propagated to LSP autocomplete (https://github.com/facebook/pyrefly/issues/1038). This diff passes along the deprecated tags in LSP auto completion.

Caveat: deprecated properties in classes are currently not tagged with deprecated tag correctly.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/1093

Reviewed By: stroxler

Differential Revision: D82451687

Pulled By: kinto0

fbshipit-source-id: 2db2560ff5e83067836e114f6857cef2fbd41022