Replace `TParamArray` with a map structure in variance inference
Summary: It's not clear to me why we were using an array to store a data structure that clearly wants good lookup behaviors. I'm guessing this comes from the fact that the original OCaml code uses OCaml array for its mutability. But we are now in Rust and can just use the mutable `SmallMap` (instead of the immutable OCaml `Map`).

Reviewed By: samwgoldman

Differential Revision: D82363091

fbshipit-source-id: 82ad2289850744d70bdaab7dc1b0887a169208e0