Key loader cache on module name and origin module path
Summary:
Upcoming changes will still require that the importing module path is used for module finding in build systems, especially with `base_module` rewriting in Buck.

Imagine the following (simplified) sourcedb output:
```
{
  "//module/import:lib": {
    "srcs": {
      "module.import.a": [
        "/module/import/a.py"
      ]
    }
  },
  "//third-party/library/target1:target1": {
    "srcs": {
      "module.import.a": [
        // this target has "module_path" = ""
        "/third-party/library/target1/module/import/a.py"
      ]
    }
  }
}
```

Due to that, we need to update our cache to key on the origin `ModulePath` too, or we might run into a situation where we keep selecting the wrong module path if two targets that depend on module names that are duplicated are open or type checked at the same time.

The approach I took to caching this is to save the `ModuePath`, since it's `Arc`'d and cheaply dupable, but has the downside of possibly (although unlikely) performing (up to 4) duplicate lookups if the same `ModulePath` is represented on the filesystem or in memory multiple times.

Alternative approaches here would be to change the `ModulePathDetails` to wrap all paths in an `Arc`, which we can use to pass around here or intern `PathBuf`s that are used in the program, since they're likely duplicated all over the place.

Reviewed By: grievejia

Differential Revision: D82123714

fbshipit-source-id: b3a894250087858625f01b76678401cc0e420020