New trait for includes
Summary:
This diff makes a new trait, `Includes` (open to renaming), which defines how file listing happens. `FilteredGlobs` are an impl of `Includes`.

The reason we're doing this is because in the near future, with build system checks, we need to be able to list files when targets are used. This just sets up for that, but the functionality actually won't be used for a while.

Reviewed By: grievejia

Differential Revision: D81510385

fbshipit-source-id: 7b604ec922d923e951de461eb6e1537d7291301a