Only store paths in `Handles`
Summary:
This diff updates `Handles` to only store `PathBuf`s within it. We do this because the extra information that's stored along with the handles will no longer necessarily be correct. We'll instead want to re-query the config for this information.

The good news is that the way `Handles` has been used so far always has a `ConfigFinder` close by, so we can change the interface for `all()` to pass in the finder there instead of on construction and updating it.

Reviewed By: grievejia

Differential Revision: D81512780

fbshipit-source-id: 72f53be7b0999de04cb58bd5f6811d7f8f520960