Make SysInfo deserializable
Summary:
I'd prefer to use `SysInfo` in the deserialized structs, since we'll be cloning stuff a lot more otherwise and wrapping things in `SysInfo` anyway. `SysInfo` is `Arc`'d, so let's just deserialize into it directly.

This diff adds a deserialization implementation that flattens the `SysInfo` into whatever struct it's being parsed from.

Reviewed By: grievejia

Differential Revision: D81627650

fbshipit-source-id: 362b1fa78a29757d62b4d038b95331236569834f