Create BuildSystem and add it to configs
Summary:
This diff creates a new enum, `BuildSystem`, which will enable users to select (and possibly configure) the build system they would like to use. It is also set as a field on `ConfigFile`, though it won't be fully usable for a bit.

In the next diff, we'll use the build system to set a sourcedb instance on the `ConfigFile`.

Reviewed By: grievejia

Differential Revision: D81510386

fbshipit-source-id: 4f5b64ff12ab802de42ff4d87292215cb5b039e8