Create target/file enum for bxl
Summary: Since build systems will be working with both files and targets, we should update the sourcedb to recognize that. This is because we probably want to be able to filter the type errors reported to the handles that are specified, regardless of if that's done by target or file.

Reviewed By: grievejia

Differential Revision: D81614215

fbshipit-source-id: d8d7167a2715bfeea50c5860e466492a8a6bb830