Move pydantic hook down
Summary:
## About this stack

See D82378108 for full details - I'm trying to unblock a change to flow-merging by first
re-structuring the scope so that there is just one map of information about names currently
tracked by the scope, and the elements of that map can distinguish between relevant cases:
- Uninitialized locals
- Initialized (or possibly-initialized) locals, including both assigned type and any narrows
- Mutable captures, including any narrows

To make this happen without a single massive change to the entire bindings
layer, I first need to encapsulate the details of the current `FlowInfo` / `StaticInfo`
layout behind higher-level APIs that more clearly describe the kinds of things other
parts of bindings are actually trying to do.

## About this sub-stack

One of the callers currently fiddling directly with FlowInfo and StaticInfo is the
class field extraction code. This sub-stack is going to refactor that logic so that
`scope` handles reading its own internals to decide what the class fields are,
and `class_def` handles only the downstream logic of tracking them in bindings.

## About this diff

The first step toward letting `scope` own its internals here is to separate reading
scope info, which is a pure function, from side-effecting logic that mutates
`fields` used in class metadata and creates bindings.

That's not possible to do cleanly right now because there's a pydantic metadata
extraction buried inside of the match on flow_info / static_info because that messes
up borrow checker rules on `&self`, so the first step is to move that down.

Reviewed By: samwgoldman

Differential Revision: D82449643

fbshipit-source-id: cedd1667e0a62ca6a6757d41ac6bf744bd0a59a7