Collect field definitions into a map
Summary:
## This Stack

We are continuing to work on moving the extraction of class field
definitions from scope data into the scope module.

This is part of a bigger refactor to decouple most of the binding logic
from details of how the scope is represented, which will help us make
scope handling simpler and more correct as well as unblock changes we
want to make to flow merging in the presence of narrows.

## This diff

The original logic for class field tried to optimize in ways that are
only really possible when we put a lot of logic all in the same big
function, by lazily iterating over the field definitions; we used
`fields` to deal with duplicates.

In order to cleanly separate finding the fields from performing mutations
while minimizing clones, I really need the ability to return owned data
and deal (when extracting fields from scope) with the fact that class body
definitions take priority over implicit definitions in methods.

I do this now: we store the `field_definitions` as a Map rather than an
iterator. This also allows us to simplify the downstream logic a bit,
since now the `field_definitions` won't contain duplicates in the first
place.

Reviewed By: samwgoldman

Differential Revision: D82449646

fbshipit-source-id: fc555b952a1720c60d3d2f7906e91995e7010d34