Split scope read vs mutations for class-body fields
Summary:
## About this stack

See D82378108 for full details - I'm trying to unblock a change to flow-merging by first
re-structuring the scope so that there is just one map of information about names currently
tracked by the scope, and the elements of that map can distinguish between relevant cases:
- Uninitialized locals
- Initialized (or possibly-initialized) locals, including both assigned type and any narrows
- Mutable captures, including any narrows

To make this happen without a single massive change to the entire bindings
layer, I first need to encapsulate the details of the current `FlowInfo` / `StaticInfo`
layout behind higher-level APIs that more clearly describe the kinds of things other
parts of bindings are actually trying to do.

## About this sub-stack

One of the callers currently fiddling directly with FlowInfo and StaticInfo is the
class field extraction code. This sub-stack is going to refactor that logic so that
`scope` handles reading its own internals to decide what the class fields are,
and `class_def` handles only the downstream logic of tracking them in bindings.

## About this diff

In order to allow scope reads to own scope internal data, we need to split the
class field extraction logic so that reads and mutating bindings happen
independently.

This diff does that for class-body-defined fields; I'll do the same for method-defined
fields next.

Reviewed By: samwgoldman

Differential Revision: D82449645

fbshipit-source-id: f5c515dbbf1b4ff68e3b2ee4920827fb9f737d72