Move class field extraction to `scope.rs`
Summary:
## About this stack

See D82378108 for full details - I'm trying to unblock a change to flow-merging by first
re-structuring the scope so that there is just one map of information about names currently
tracked by the scope, and the elements of that map can distinguish between relevant cases:
- Uninitialized locals
- Initialized (or possibly-initialized) locals, including both assigned type and any narrows
- Mutable captures, including any narrows

To make this happen without a single massive change to the entire bindings
layer, I first need to encapsulate the details of the current `FlowInfo` / `StaticInfo`
layout behind higher-level APIs that more clearly describe the kinds of things other
parts of bindings are actually trying to do.

## About this sub-stack

One of the callers currently fiddling directly with FlowInfo and StaticInfo is the
class field extraction code. This sub-stack is going to refactor that logic so that
`scope` handles reading its own internals to decide what the class fields are,
and `class_def` handles only the downstream logic of tracking them in bindings.

## About this diff

This diff completes the sub-stack - there is now a helper function on `Scope` that will
consume a class body scope and return all of the class field definitions.

Reviewed By: samwgoldman

Differential Revision: D82449644

fbshipit-source-id: 95f5881532ff311c280831cf735b956114a3645e