Move class_scope up, fields down
Summary:
## This stack

I'm continuing to work on pulling class field extraction from scope into
a scope helper, which is part of a bigger effort to decouple most of Pyrefly's
bindings layer from trying to directly fiddle with `flow_info` and `static_info`.

The ultimate goal is to rework how scope internals work to be simpler and more
correct, and unblock changes we want to make to how flow merging works in the
presence of narrows.

## This diff

Now that we've separated pulling data from scope versus mutating `fields`
and defining bindings, it's easier to see the path forward if we pull
the `class_scope` extraction up so that it's more obvious what assumptions
we are making about `last_scope`, and push the `fields` initialization down
so that it's more obvious that we don't use it until after we've found
all the fields.

Reviewed By: samwgoldman

Differential Revision: D82449648

fbshipit-source-id: 0bfb02b41d473323cefee6ea4a61ce688717801b