Make finishing a class a higher-level scopes operation
Summary:
In diff review of D82449644, we noted that it would be nicer if the popping
of scopes when we finish a class lived next to the panicking code that
requires the class body to have the correct scope kind.

One option would be to leave that logic in `class.rs`, but it's hard to do
that without coupling to scope internals. The appraoch taken here is instead
to let the scope stack own the entire process of finishing a class, which
is also consistent with the pattern taken in some other cases like a module
where we have a "finish" operation that also asserts invariants.

Reviewed By: samwgoldman

Differential Revision: D82466237

fbshipit-source-id: 5b2f3c7420cfe915fe231fb01039279f69e3e1a3