Decouple export key building from scope internals
Summary:
## About this stack

See D82378108 for full details - I'm trying to unblock a change to flow-merging by first
re-structuring the scope so that there is just one map of information about names currently
tracked by the scope, and the elements of that map can distinguish between relevant cases:
- Uninitialized locals
- Initialized (or possibly-initialized) locals, including both assigned type and any narrows
- Mutable captures, including any narrows

To make this happen without a single massive change to the entire bindings
layer, I first need to encapsulate the details of the current `FlowInfo` / `StaticInfo`
layout behind higher-level APIs that more clearly describe the kinds of things other
parts of bindings are actually trying to do.

## About this diff

This diff uses a structured API to communicate everything needed
from the top-level scope to the step where we set up export bindings,
without the exports directly relying on details of flow_info and
static_info.

Reviewed By: samwgoldman

Differential Revision: D82487517

fbshipit-source-id: 6cc0f84cd8e76574a0240b4eedc528dc3d079385