Simplify flow-style-related logic
Summary:
This diff continues my work trying to simplify the API exposed by scope
so that less of bindings couples to implementation details that have to
change if we represent narrows and mutable captures differently.

The flow-style hooks we had were pretty complicated, and weren't really
necessary - if we just use whether the current name lookup is a static
type context to prune uninitialized error messages, we don't need to be
passing extra flags around to mangle the way flow style is propagated.

Reviewed By: samwgoldman

Differential Revision: D82487544

fbshipit-source-id: 4809134207bc70851766371d4e01ed00a3feebd3