Let scope handle finding mergable import key
Summary:
## About this stack

See D82378108 for full details - I'm trying to unblock a change to flow-merging by first
re-structuring the scope so that there is just one map of information about names currently
tracked by the scope, and the elements of that map can distinguish between relevant cases:
- Uninitialized locals
- Initialized (or possibly-initialized) locals, including both assigned type and any narrows
- Mutable captures, including any narrows

To make this happen without a single massive change to the entire bindings
layer, I first need to encapsulate the details of the current `FlowInfo` / `StaticInfo`
layout behind higher-level APIs that more clearly describe the kinds of things other
parts of bindings are actually trying to do.

## About this diff

We don't really want `stmt.rs` directly working with the `FlowInfo`, which couples
that logic to the existing layout of Scope (and therefore gets in the way of storing
narrows differently, as well as other bugfixes).

Instead, let's let a helper on `Scope` handle this. It's also nice that the helper
gives us a good place for a doc comment, since I'm not sure anyone who's less
than an expert at Python would immediately know what's going on here.

Reviewed By: samwgoldman

Differential Revision: D82487508

fbshipit-source-id: 3fac317bb6075636b4401c2047cb2e6f522e0b55