Less verbose names
Summary:
I sometimes like for names to emphasize when a function does something that
might not happen (in practice, many such functions do something that almost
never happens - they typically catch some edge case that isn't the normal case,
for example the function for recording `self` attribute assignments is called
on *every* attribute assignment).

But we already have a convention of naming functions that normally do
nothing for what they might do, so that the code reads as if every special
case happens all of the time.

That's not a bad thing - the names of things get pretty verbose if we always
reflect "possible" and "maybe" in function and variable names, and with inlay
hints it's often at least somewhat obvious from the type signature.

But as a result, I think the names I picked for first-use lookup handling are
inconsistent with Pyrefly naming conventions - we should just name them as
if the operations always occur, which is more concise.

Reviewed By: grievejia, ndmitchell

Differential Revision: D82498501

fbshipit-source-id: 9e0622f5ce161c872b9d11bdbde8f472319c8050