Clean up unneeded `default` in bindings.rs
Summary:
In the previous diffs, I noted that we were creating a bunch of
unnecessary `Default` nodes, which wasted heap allocations and also
forced a bunch of interfaces.

Now that they aren't needed, we no longer need `bind_name` and `bind_current`
to be returing a `default`. This is great! I always found it confusing
to keep track of which bind functions returned a default and what it was
actually used for.

Note that the `scope` code could still be simplified, I'll do that in the
next diff.

Reviewed By: samwgoldman

Differential Revision: D82487515

fbshipit-source-id: 1268ceb74bfb1acf35f238499fb52523b967120b