Stop making `Default` bindings in target.rs
Summary:
Now that we break cycles deterministically and we use an ordering where
`Key`s are sorted by `Idx`, a Phi is going to get priority over any bindings
that come below (because the speculative insert always happens earlier in
our AST traversal).

As a result, wrapping up name assignments in a `Default` node has no
benefitical effect; the only thing it ever did was try to force an earlier
binding, but the Phi is already going to do that anyway.

Not only is there no benefit, but there are real downsides:
- There's some perf cost because we are making bindings bigger and
  allocating more boxes on the heap, plus in some edge cases we would
  unnecessarily solve extra bindings (especially in return type inference)
  that weren't actually needed.
- There's significant readability cost, because I don't think anyone really
  understands default very well so at this point it was just a magic
  incantation that was scary to modify even though it no longer matters.
- It causes architectural problems - we're returning the default on
  flow updates, which never really made sense conceptually (I think it
  was originally an optimization to avoid two hashed lookups?) and is
  getting in the way of refactors as well as coupling unrelated code to
  a loop handling paradigm that we want to get rid of.

This diff turns off the use of default, I'll clean up some helper
functions in stacked diffs.

Reviewed By: samwgoldman

Differential Revision: D82487510

fbshipit-source-id: 6bd18b3606285a1038c803838ada9eb399f537e3