Stop making `Default` in `stmt.rs`
Summary:
By the same logic as the previous diff, we don't need to be creating
`Default` nodes that force extra keys in any of the `stmt.rs`
assignment handlers - these days, any recursive Phi in a loop will
break its own cycle, which means the only place we need the `Default`
is on that recursive Phi (and that's what we have - `flow_ops.rs`
always uses a `Default(_, Phi(..))` for the loop phi binding).

These other cases can all be simplified.

Reviewed By: samwgoldman

Differential Revision: D82487513

fbshipit-source-id: 97251c320a78df92ec5654710571b042936b0a8d