Track typevar instantiation errors
Summary:
When we instantiate a quantified var, we also need to ensure the instantiation is compatible with the typevar bounds.

Currently, we greedily instantiate a var the first time it appears in a subset constraint. Eventually, we will accumulate bounds and solve later in finish_quantified. For now, we just record any error at the moment of instantiation, and look it up in finish_quantified.

We should look at this result whever we instantiate, but for now we only look during subset, when we instantiate a forall lower bound. This is all we need to support the next diff, which implements generic subtyping.

Reviewed By: stroxler

Differential Revision: D82493517

fbshipit-source-id: e2716683d5ff05f9579470340ef51bbd72794c26