Split scope reading from mutations in lookup logic
Summary:
## About this stack

See D82378108 for full details - I'm trying to unblock a change to flow-merging by first
re-structuring the scope so that there is just one map of information about names currently
tracked by the scope, and the elements of that map can distinguish between relevant cases:
- Uninitialized locals
- Initialized (or possibly-initialized) locals, including both assigned type and any narrows
- Mutable captures, including any narrows

To make this happen without a single massive change to the entire bindings
layer, I first need to encapsulate the details of the current `FlowInfo` / `StaticInfo`
layout behind higher-level APIs that more clearly describe the kinds of things other
parts of bindings are actually trying to do.

## About this sub-stack

In this sub-stack, I make the scope stack own how to look up the currently bound key
for a name, by cleanly separating that operation (which is a pure function and depends only
on scope semantics) from the mutations involved in intercepting first usage (which is a binding
table concern, not something modeled in scope) and mutating the tables (again, a binding table
concern).

## About this diff

This diff kicks off the sub-stack by factoring most of the original implementation of
`get_bond_key_and_first_use_at_name` into a new, pure function `key_for_read_of_name`
(not how `&self` is not mutable in the newly-split-out function).

At the moment I use an Either type inside a `Result` which is messy but allows the raw
split to be done in a minimal diff; I'll change it to an equivalent but more descriptive
type later after I move the logic to the scope module.

Reviewed By: grievejia

Differential Revision: D82498638

fbshipit-source-id: f01109f037372da30d2a53c5308eec7c4cfd5a29