Move scope-only logic to scope module
Summary:
## About this stack

See D82378108 for full details - I'm trying to unblock a change to flow-merging by first
re-structuring the scope so that there is just one map of information about names currently
tracked by the scope, and the elements of that map can distinguish between relevant cases:
- Uninitialized locals
- Initialized (or possibly-initialized) locals, including both assigned type and any narrows
- Mutable captures, including any narrows

To make this happen without a single massive change to the entire bindings
layer, I first need to encapsulate the details of the current `FlowInfo` / `StaticInfo`
layout behind higher-level APIs that more clearly describe the kinds of things other
parts of bindings are actually trying to do.

## About this sub-stack

In this sub-stack, I make the scope stack own how to look up the currently bound key
for a name, by cleanly separating that operation (which is a pure function and depends only
on scope semantics) from the mutations involved in intercepting first usage (which is a binding
table concern, n

## About this diff

This diff moves the split-out pure scope-only lookup function into `scope.rs` and makes
no other changes.

Reviewed By: grievejia

Differential Revision: D82498634

fbshipit-source-id: 5b82456f41c4798f34c405c13de1504ebdffceaa