Create a structured api for name write information
Summary:
## About this stack

See D82378108 for full details - I'm trying to unblock a change to flow-merging by first
re-structuring the scope so that there is just one map of information about names currently
tracked by the scope, and the elements of that map can distinguish between relevant cases:
- Uninitialized locals
- Initialized (or possibly-initialized) locals, including both assigned type and any narrows
- Mutable captures, including any narrows

To make this happen without a single massive change to the entire bindings
layer, I first need to encapsulate the details of the current `FlowInfo` / `StaticInfo`
layout behind higher-level APIs that more clearly describe the kinds of things other
parts of bindings are actually trying to do.

## About this diff

To perform a write to a name, we need to do a scope lookup that returns
different information than reads need. The things we need for a write are:
- The annotation (if any) associated with this name, which will be used for
  both contextual typing and to verify that we don't violate it as an upper
  bound
- Enough knowledge to handle recording the write in an `Anywhere` binding
  when necessary. We need the `range` used on the `Anywhere` for this, but
  we only need it when `definitions.rs` identified multiple assignments,
  otherwise we skip it as an optimization.

This diff creates a datatype to get this information in a structured
way, and migrates `bind_name` to an API instead of accessing scope
internals directly.

Reviewed By: grievejia

Differential Revision: D82548822

fbshipit-source-id: 510da3a04e30deb56bbb9030c5dfc98a986e601f